move series and standfirst into shared
